+++ ProPPle/app/src/main/java/com/example/propple/viewModel/clientePrestador/FormularioReservaViewModel.kt
@@ -1,13 +1,10 @@
 package com.ort.casodeusotest.viewModel
 
-import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.propple.api.RetrofitHelper
 import com.example.propple.api.Transacciones.FormalizarTransaccion
-import com.example.propple.api.Transacciones.VentasPro
 import com.example.propple.api.interfaces.Transacciones
-import com.example.propple.shared.ProPPle
 import com.example.propple.shared.ProPPle.Companion.prefs
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -17,12 +14,14 @@ import retrofit2.Response
 class FormularioReservaViewModel : ViewModel() {
     var status= MutableLiveData<Boolean>()
 
-    fun enviarFormulario(date_of_work: String,
-                         id_transaccion: Int,
-                         location: String,
-                         location_latitud: Double,
-                         location_longitud: Double,
-                         presupuesto: Int) {
+    fun enviarFormulario(
+        date_of_work: String,
+        id_transaccion: Int,
+        location: String,
+        location_latitud: Double,
+        location_longitud: Double,
+        presupuesto: Double
+    ) {
         val formulario = FormalizarTransaccion(date_of_work,
         id_transaccion,
         location,
